save images to user folder,delete images, delete full image history by id

+++ context/HistoryContextProvider.js
@@ -2,9 +2,11 @@ import { createContext, useContext, useEffect, useMemo, useReducer } from "react
 import { db } from "../firebaseConfig";
 import { get, ref, child, update } from "firebase/database";
 import { useAuthContext } from "./AuthContextProvider";
+import { fsAPI } from "../lib/fsAPI";
 
 const dbRef = ref(db);
 const DB_CHAT_PATH = process.env.EXPO_PUBLIC_DB_CHAT_PATH;
+const DB_IMAGE_PATH = process.env.EXPO_PUBLIC_DB_IMAGE_PATH;
 const DB_USER_PATH = process.env.EXPO_PUBLIC_DB_USER_PATH;
 
 const HistoryContext = createContext();
@@ -125,6 +127,18 @@ const initialState = {
 
 const reducer = (prevState, action) => {
     switch (action.type) {
+
+        case 'GET-HISTORY-INDEXES': {
+
+            return {
+                ...prevState,
+                ['chatHistory']: {
+                    ...prevState['chatHistory'],
+                    historyIndexes: action.payload
+                }
+            }
+        }
+
         case 'SET-CHAT-HISTORY-ID':
             return {
                 ...prevState,
@@ -135,15 +149,7 @@ const reducer = (prevState, action) => {
 
             }
 
-        case 'GET-HISTORY-INDEXES': {
-            return {
-                ...prevState,
-                ['chatHistory']: {
-                    ...prevState['chatHistory'],
-                    historyIndexes: action.payload
-                }
-            }
-        }
+
         case 'GET-HISTORY-BY-PROVIDED-ID':
             return {
                 ...prevState,
@@ -151,6 +157,18 @@ const reducer = (prevState, action) => {
                     ...prevState['chatHistory'],
                     ['history']: {
                         ...prevState['chatHistory']['history'],
+                        [action.payload.id]: action.payload.data
+
+                    }
+                }
+            }
+        case 'GET-IMAGES-HISTORY-BY-PROVIDED-ID':
+            return {
+                ...prevState,
+                ['imagesHistory']: {
+                    ...prevState['imagesHistory'],
+                    ['history']: {
+                        ...prevState['imagesHistory']['history'],
                         [action.payload.id]: action.payload.data
 
                     }
@@ -221,6 +239,15 @@ const reducer = (prevState, action) => {
 
 
         // IMAGES related
+
+        case 'GET-IMAGES-HISTORY-INDEXES':
+            return {
+                ...prevState,
+                ['imagesHistory']: {
+                    ...prevState['imagesHistory'],
+                    historyIndexes: action.payload
+                }
+            }
         case 'SET-IMAGES-HISTORY-ID':
             return {
                 ...prevState,
@@ -230,7 +257,7 @@ const reducer = (prevState, action) => {
                 }
             }
         case 'ADD-IMAGE-TO-HISTORY':
-            console.log('data in reducer:', action.payload)
+
             if (prevState['imagesHistory']['history'][action.payload.historyId]) {
                 return {
                     ...prevState,
@@ -392,10 +419,24 @@ const HistoryContextProvider = ({ children }) => {
                     console.log('error while delete chat item')
                 });
         },
-        setImagesHistoryId: (value) => {
-            if (value) {
+        setImagesHistoryId: (valueId) => {
+
+            if (valueId) {
+                // check if 'valueId' exists at local state history, if NOT - request history data with such 'valueId' from server  
+                if (historyState.imagesHistory.history && !historyState.imagesHistory.history[valueId]) {
+
+                    get(child(dbRef, DB_IMAGE_PATH + userId + '/' + valueId))
+                        .then(snapshot => {
+                            if (snapshot.exists()) {
+                                const server_data = snapshot.val();
+                                dispatch({ type: 'GET-IMAGES-HISTORY-BY-PROVIDED-ID', payload: { id: valueId, data: server_data } })
+                            } else {
+                                console.log("No data available");
+                            }
+                        });
+                }
                 dispatch(
-                    { type: 'SET-IMAGES-HISTORY-ID', payload: value }
+                    { type: 'SET-IMAGES-HISTORY-ID', payload: valueId }
                 )
             } else {
                 dispatch(
@@ -405,13 +446,68 @@ const HistoryContextProvider = ({ children }) => {
         },
 
         addImagesHistoryItem: ({ historyId, data }) => {
-            dispatch({ type: 'ADD-IMAGE-TO-HISTORY', payload: { historyId, value: data } })
+
+            const updates = {};
+
+            // send data to server
+            if (historyState.imagesHistory.history && historyState.imagesHistory.history[historyId]) {
+                // if history of current chat exists
+                updates[DB_IMAGE_PATH + userId + '/' + historyId] = [...historyState.imagesHistory.history[historyId], data];
+
+            } else {
+                //if no history  
+                updates[DB_IMAGE_PATH + userId + '/' + historyId] = [data];
+                updates[DB_USER_PATH + userId + '/images'] = { ...historyState.imagesHistory.historyIndexes, [historyId]: data.title }
+            }
+            update(dbRef, updates)
+                .then(() => {
+                    // update local state 
+                    dispatch({ type: 'ADD-IMAGE-TO-HISTORY', payload: { historyId, value: data } });
+                })
+                .catch((error) => console.log('error while add images history item ', error));
         },
+
         deleteImageItem: ({ historyId, imageSource }) => {
-            dispatch({ type: 'DELETE-IMAGE-ITEM', payload: { historyId, imageSource } })
+            const updates = {};
+            let newHistory = historyState.imagesHistory.history[historyId].filter(item => item.source !== imageSource);
+
+            updates[DB_IMAGE_PATH + userId + '/' + historyId] = newHistory;
+            update(dbRef, updates)
+                .then(() => {
+                    // remove file from user folder
+                    fsAPI.deleteFileInUserFolder(imageSource).then((res) => {
+                        if (res.status == 'Success') {
+                            dispatch({ type: 'DELETE-IMAGE-ITEM', payload: { historyId, imageSource } })
+                        }
+                    })
+                })
+                .catch((error) => {
+                    //  delete operation failed...
+                    console.log(`Error while delete a single image item: `, error)
+                });
+
         },
-        deleteImagesHistoryItem: (historyId) => {
-            dispatch({ type: 'DELETE-IMAGES-HISTORY-ITEM', payload: historyId })
+        deleteImagesHistoryItem: async (historyId) => {
+            const updates = {};
+            const { [historyId]: removedData, ...restDataIndexes } = historyState.imagesHistory.historyIndexes;
+
+            updates[DB_USER_PATH + userId + '/images'] = { ...restDataIndexes };
+            updates[DB_IMAGE_PATH + userId + '/' + historyId] = null;
+
+            update(dbRef, updates)
+                .then(() => {
+
+                    fsAPI.deleteFilesFromAppStorage(historyId)
+                        .then(res => {
+                            if (res.status == 'Success') {
+                                dispatch({ type: 'DELETE-IMAGES-HISTORY-ITEM', payload: historyId })
+                            }
+                        });
+                })
+                .catch((error) => {
+                    //  writing failed...
+                    console.log('Error while delete images history item: ', error)
+                });
         }
 
 
@@ -429,16 +525,30 @@ const HistoryContextProvider = ({ children }) => {
         //PROD
         //
         //get data from online db
-        if (userId) {
-            get(child(dbRef, DB_USER_PATH + userId + '/chats'))
+
+        const checkIndexes = async (userId) => {
+            await get(child(dbRef, DB_USER_PATH + userId + '/chats'))
                 .then(snapshot => {
                     if (snapshot.exists()) {
                         const server_data = snapshot.val();
                         dispatch({ type: 'GET-HISTORY-INDEXES', payload: server_data });
                     } else {
-                        console.log("No data available");
+                        // console.log("No chats indexes data available");
                     }
                 });
+
+            await get(child(dbRef, DB_USER_PATH + userId + '/images'))
+                .then(snapshot => {
+                    if (snapshot.exists()) {
+                        const server_data = snapshot.val();
+                        dispatch({ type: 'GET-IMAGES-HISTORY-INDEXES', payload: server_data });
+                    } else {
+                        //  console.log("No images indexes data available");
+                    }
+                });
+        }
+        if (userId) {
+            checkIndexes(userId)
         }
     }, [userId])
